tasks : more homogeneous handling of interface ids
Previsoulsy an interface was identified by a tuple containing its name and its anchor (task_id and interfaces names). This was differing from task in that only the task name was used. Now the full id (with root package) is used and the id is a string ':' separated of all the ids (task and interfaces)