tasks: fix linspace interface
The last fix to linspace interface introduced some random issues with handling of the last point due to floating point rounding errors. This patch revert to using linspace for the array generation without losing the rounding of values (now based also on the number of digits in start and stop in addition to step), nor the handling of stop not matching the given start and step.