measurement: fix accessing runtime dependencies
Guarantee that even if no dependency of a given type could be collected, the corresponding section will still exist in the dict returned by MeasurementDependencies.get_runtime_dependencies.

Use this guarantee to simplify/fix the code of InstrTask (previously we would have gotten a NameError if the profile section was absent, bad tests...).